requirements.txt
dagit==1.3.11
dagster==1.3.11
dagster-graphql==1.3.11
dagster-postgres==0.19.11
PyGithub==1.59.0